tag controller, with tagCard() method

+++ grids/src/grids/web/UserController.java
@@ -23,7 +23,7 @@ public class UserController {
 	}
 	
 	@RequestMapping("/info/{id}")
-	public ModelAndView userInfo(@PathVariable int id) {
+	public ModelAndView userInfo(@PathVariable long id) {
 		return null;
 	}
 }
